# frontend/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Annotate Tool",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/quality-batches": {
      "post": {
        "summary": "Create Quality Batch",
        "operationId": "create_quality_batch_api_v1_quality_batches_post",
        "parameters": [
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityBatchCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityBatchDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Quality Batches",
        "operationId": "list_quality_batches_api_v1_quality_batches_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "name": "package_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Package Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QualityBatchOut"
                  },
                  "title": "Response List Quality Batches Api V1 Quality Batches Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quality-batches/{batch_id}": {
      "get": {
        "summary": "Get Quality Batch",
        "operationId": "get_quality_batch_api_v1_quality_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityBatchDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-items/{item_id}/quality-history": {
      "get": {
        "summary": "Quality History",
        "operationId": "quality_history_api_v1_work_items__item_id__quality_history_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QualityCheckOut"
                  },
                  "title": "Response Quality History Api V1 Work Items  Item Id  Quality History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-items/{item_id}/quality-check": {
      "post": {
        "summary": "Quality Check",
        "operationId": "quality_check_api_v1_work_items__item_id__quality_check_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quality-batches/{batch_id}/items/{item_id}/check": {
      "post": {
        "summary": "Check Quality Batch Item",
        "operationId": "check_quality_batch_item_api_v1_quality_batches__batch_id__items__item_id__check_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "annotate_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Annotate Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ItemStatus": {
        "type": "string",
        "enum": [
          "available",
          "annotation_assigned",
          "annotating",
          "review_pending",
          "review_assigned",
          "reviewing",
          "changes_requested",
          "completed",
          "cancelled"
        ],
        "title": "ItemStatus"
      },
...
      "QaStatus": {
        "type": "string",
        "enum": [
          "unchecked",
          "passed",
          "rejected"
        ],
        "title": "QaStatus"
      },
      "QualityBatchCreate": {
        "properties": {
          "package_id": {
            "type": "string",
            "title": "Package Id"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "all",
              "ratio",
              "count"
            ],
            "title": "Mode",
            "default": "ratio"
          },
          "percent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Percent",
            "default": 10
          },
          "count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Count",
            "default": 20
          },
          "seed": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Seed",
            "default": "quality"
          },
          "only_unchecked": {
            "type": "boolean",
            "title": "Only Unchecked",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "package_id"
        ],
        "title": "QualityBatchCreate"
      },
      "QualityBatchDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "package_id": {
            "type": "string",
            "title": "Package Id"
          },
          "created_by_id": {
            "type": "string",
            "title": "Created By Id"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "sample_percent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Percent"
          },
          "sample_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Count"
          },
          "seed": {
            "type": "string",
            "title": "Seed"
          },
          "only_unchecked": {
            "type": "boolean",
            "title": "Only Unchecked"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "total_samples": {
            "type": "integer",
            "title": "Total Samples"
          },
          "checked_samples": {
            "type": "integer",
            "title": "Checked Samples"
          },
          "passed_samples": {
            "type": "integer",
            "title": "Passed Samples"
          },
          "rejected_samples": {
            "type": "integer",
            "title": "Rejected Samples"
          },
          "samples": {
            "items": {
              "$ref": "#/components/schemas/QualitySampleOut"
            },
            "type": "array",
            "title": "Samples"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "package_id",
          "created_by_id",
          "assignee_id",
          "mode",
          "sample_percent",
          "sample_count",
          "seed",
          "only_unchecked",
          "status",
          "created_at",
          "completed_at",
          "total_samples",
          "checked_samples",
          "passed_samples",
          "rejected_samples",
          "samples"
        ],
        "title": "QualityBatchDetailOut"
      },
      "QualityBatchOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "package_id": {
            "type": "string",
            "title": "Package Id"
          },
          "created_by_id": {
            "type": "string",
            "title": "Created By Id"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "sample_percent": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Percent"
          },
          "sample_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Count"
          },
          "seed": {
            "type": "string",
            "title": "Seed"
          },
          "only_unchecked": {
            "type": "boolean",
            "title": "Only Unchecked"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "total_samples": {
            "type": "integer",
            "title": "Total Samples"
          },
          "checked_samples": {
            "type": "integer",
            "title": "Checked Samples"
          },
          "passed_samples": {
            "type": "integer",
            "title": "Passed Samples"
          },
          "rejected_samples": {
            "type": "integer",
            "title": "Rejected Samples"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "package_id",
          "created_by_id",
          "assignee_id",
          "mode",
          "sample_percent",
          "sample_count",
          "seed",
          "only_unchecked",
          "status",
          "created_at",
          "completed_at",
          "total_samples",
          "checked_samples",
          "passed_samples",
          "rejected_samples"
        ],
        "title": "QualityBatchOut"
      },
      "QualityCheckOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "task_item_id": {
            "type": "string",
            "title": "Task Item Id"
          },
          "reviewer_id": {
            "type": "string",
            "title": "Reviewer Id"
          },
          "result": {
            "$ref": "#/components/schemas/QaStatus"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "revision_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Id"
          },
          "revision_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Version"
          },
          "revision_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Hash"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "task_item_id",
          "reviewer_id",
          "result",
          "comment",
          "revision_id",
          "revision_version",
          "revision_hash",
          "created_at"
        ],
        "title": "QualityCheckOut"
      },
      "QualityInput": {
        "properties": {
          "result": {
            "$ref": "#/components/schemas/QaStatus"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "QualityInput"
      },
      "QualitySampleOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "task_item_id": {
            "type": "string",
            "title": "Task Item Id"
          },
          "sample_order": {
            "type": "integer",
            "title": "Sample Order"
          },
          "item": {
            "$ref": "#/components/schemas/TaskItemOut"
          },
          "latest_check": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualityCheckOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "task_item_id",
          "sample_order",
          "item"
        ],
        "title": "QualitySampleOut"
      },
...
      "TaskItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "package_id": {
            "type": "string",
            "title": "Package Id"
          },
          "episode_id": {
            "type": "string",
            "title": "Episode Id"
          },
          "claim_order": {
            "type": "integer",
            "title": "Claim Order"
          },
          "status": {
            "$ref": "#/components/schemas/ItemStatus"
          },
          "annotator_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotator Id"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "qa_status": {
            "$ref": "#/components/schemas/QaStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "package_id",
          "episode_id",
          "claim_order",
          "status",
          "annotator_id",
          "reviewer_id",
          "qa_status",
          "updated_at"
        ],
        "title": "TaskItemOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}